Extract vscalar tests
Keep callback coverage intact while leaving room for later module growth.
